requirements.txt
dnspython==2.6.1
httpx==0.27.2
httpx[http2]==0.27.2
jinja2==3.1.4
python-whois==0.9.5
beautifulsoup4==4.12.3
favicon==0.7.0
idna==3.6
rich==13.7.1
certifi>=2024.7.4
chardet>=5.2.0
urllib3>=2.2.2
